Moved executables to scripts directory
